+++ LaryngektomieWeb/src/main/java/cz/laryngektomie/controller/AdminController.java
@@ -1,4 +1,4 @@
-package cz.laryngektomie.controler;
+package cz.laryngektomie.controller;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.access.prepost.PreAuthorize;
